Higher thread level parallelism improvement
1) Changed the code to be less lock contentive when the thread count is
greater than 10.
2) This change in code has not only improved thread parallelism and
speed in higher thread count applications, it has also increased the
speed of the program from 60 to 53 seconds for the Boosh.avi file
processing.
3) One thing to note is the algorithm flourishes off finding similar
patterns.  The more variability of unique bytes in a file will yield
more lock contention while say a binary file that has 1s and 0s will run
very good with many threads.

+++ BaseCodeCompletelyParallel/TypeDefines.h
@@ -13,8 +13,7 @@ using namespace std;
 #if GREATERTHAN4GB
 typedef unsigned long long PListType;
 #else
-//typedef unsigned long PListType ;
-typedef unsigned int PListType ;
+typedef unsigned long PListType ;
 #endif
 
 #if INTEGERS
